--- src/main/java/br/edu/ifpb/mt/dac/dao/PersistenciaDacException.java
@@ -1,17 +1,0 @@
-package br.edu.ifpb.mt.dac.dao;
-
-import br.edu.ifpb.mt.dac.exceptions.DacException;
-
-public class PersistenciaDacException extends DacException {
-
-	private static final long serialVersionUID = 7159282553688713660L;
-
-	public PersistenciaDacException(String message) {
-		super(message);
-	}
-
-	public PersistenciaDacException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-}
