--- src/main/java/br/edu/ifpb/mt/dac/dtos/livro/LivroUpdateDTO.java
@@ -1,17 +1,0 @@
-package br.edu.ifpb.mt.dac.dtos.livro;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class LivroUpdateDTO {
-    private String nome;
-    private String autor;
-    private String genero;
-    private LocalDate localDate;
-}
