--- src/main/java/br/edu/ifpb/mt/dac/enuns/GeneroLivro.java
@@ -1,9 +1,0 @@
-package br.edu.ifpb.mt.dac.enuns;
-
-public enum GeneroLivro {
-    ROMANCE,
-    TERROR,
-    SUSPENSE,
-    FICCAO_CIENTIFICA,
-    FANTASIA
-}
