+++ src/main/java/projetobdII/biblioteca/dao/DAO.java
@@ -1,0 +1,24 @@
+package projetobdII.biblioteca.dao;
+
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
+
+public abstract class DAO {
+
+	static EntityManagerFactory emf;
+
+	static {
+		emf = Persistence.createEntityManagerFactory("dao");
+	}
+
+	protected EntityManager getEntityManager() {
+		return emf.createEntityManager();
+	}
+
+	public void close() {
+		if (emf.isOpen()) {
+			emf.close();
+		}
+	}
+}
