+++ src/main/java/projetobdII/biblioteca/dao/PersistenciaDacException.java
@@ -1,0 +1,17 @@
+package projetobdII.biblioteca.dao;
+
+import projetobdII.biblioteca.exceptions.DacException;
+
+public class PersistenciaDacException extends DacException {
+
+	private static final long serialVersionUID = 7159282553688713660L;
+
+	public PersistenciaDacException(String message) {
+		super(message);
+	}
+
+	public PersistenciaDacException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+}
