+++ src/main/java/projetobdII/biblioteca/dtos/livro/LivroDeleteDTO.java
@@ -1,0 +1,18 @@
+package projetobdII.biblioteca.dtos.livro;
+
+import java.time.LocalDate;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class LivroDeleteDTO {
+    
+    private String nome;
+    private String autor;
+    private String genero;
+    private LocalDate localDate;
+
+}
