+++ src/main/java/projetobdII/biblioteca/enuns/GeneroLivro.java
@@ -1,0 +1,9 @@
+package projetobdII.biblioteca.enuns;
+
+public enum GeneroLivro {
+    ROMANCE,
+    TERROR,
+    SUSPENSE,
+    FICCAO_CIENTIFICA,
+    FANTASIA
+}
