+++ src/main/java/projetobdII/biblioteca/exceptions/RegraNegocioException.java
@@ -1,0 +1,12 @@
+package projetobdII.biblioteca.exceptions;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@AllArgsConstructor
+@Data
+public class RegraNegocioException extends Exception{
+    public RegraNegocioException(String message){
+        super(message);
+    }
+}
