+++ src/main/java/projetobdII/biblioteca/mocks/MainSave.java
@@ -1,0 +1,32 @@
+package projetobdII.biblioteca.mocks;
+
+import projetobdII.biblioteca.dao.LivroDAO;
+import projetobdII.biblioteca.enuns.GeneroLivro;
+import projetobdII.biblioteca.entities.LivroEntity;
+import projetobdII.biblioteca.exceptions.DacException;
+
+import java.time.LocalDate;
+
+public class MainSave {
+
+	public static void main(String[] args) throws DacException {
+		LivroDAO dao = new LivroDAO();
+		try {
+			LivroEntity livroEntity = new LivroEntity();
+
+			livroEntity.setNome("teste2");
+			livroEntity.setGenero(GeneroLivro.FICCAO_CIENTIFICA);
+			livroEntity.setAutor("carlos");
+			livroEntity.setDataDePublicacao(LocalDate.now());
+
+			System.out.println(livroEntity);
+			
+			dao.save(livroEntity);
+
+			System.out.println(livroEntity);
+		} finally {
+			dao.close();
+		}
+	}
+
+}
