+++ src/main/resources/META-INF/persistence.xml
@@ -7,7 +7,7 @@
 		<provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
 		<!-- <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider> -->
 
-		<class>br.edu.ifpb.mt.dac.entities.LivroEntity</class>
+		<class>projetobdII.biblioteca.entities.LivroEntity</class>
 
 		<properties>
